move modules into core, services, and api/v1/; tests passing locally
