+++ refiner/requirements.txt
@@ -1,7 +1,6 @@
 fastapi>=0.109.1
 httpx
 lxml
-pathlib
 pydantic-settings
 pydantic>=2.0.0
 python-multipart
